moving page reducer to the compose-core module
