+++ features/messenger/src/main/kotlin/app/dapk/st/messenger/gallery/ImageGalleryScreen.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.unit.sp
 import app.dapk.st.core.Lce
 import app.dapk.st.core.LifecycleEffect
 import app.dapk.st.core.components.CenteredLoading
+import app.dapk.st.core.page.PageAction
 import app.dapk.st.design.components.GenericError
 import app.dapk.st.design.components.Spider
 import app.dapk.st.design.components.SpiderPage
